requirements.txt
fastapi
uvicorn
pdfplumber
camelot-py[cv]
pdfminer.six
python-multipart
python-dotenv
